new btn-group--stacked molecule
adds a new CSS molecule for stacking buttons vertically and stretching
their widths to match the width of the container. ideally they should
only stretch to the natural width of the widest button under all
circumstances; for now this is only the case when the btn-group--stacked
container is a flex-child.